pyproject.toml
[project]
name = "5-workshop-restapi-fastapi-deploys"
version = "0.1.0"
description = ""
authors = [
    {name = "Paulo Cardoso",email = "cardosopaulohenrique91@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "requests (>=2.32.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
psycopg2-binary = {source = "pypi"}
